requirements.txt
Flask
Flask-Bootstrap
Pillow
requests
opencv-contrib-python
os 
sys
time
random
argparse
werkzeug